Make all rolls valid for the player
For example, counter_slash.dart had "Will you keep your balance" as a rollReason. That was stupid. Now we check for success of what the player actually wants to do, that is _hitting_ the opponent.

This means we need to make special version of some moves so that the player’s version will pre-determine the result of defense moves. The predetermination is added to the defense situation, and — by convention — actions have to follow the predetermination.