Separate Action’s name from command
`name` is the internal String representing the class (kind of like runtimeType, but guaranteed to be the same even when compiled to JavaScript). `command` (previously known as `name`) is the string presented to the player as command.

For example, a TakeExitAction has name of “TakeExitAction” and command of “Go to the kitchen”.

This fixes the issue of actors standing up right after being knocked down even despite the logic in StandUp and Scramble. The logic depended upon runtimeType of previous actions being exactly “Pound” or “SweepOffFeet” (but after `dart2js —mode release` they were something like “dgf”).