Add note about info to game
Displays “Debug version from Wed, March 15, 2017.” at the start of the game. This is to be manually updated for every release.

The note helps to understand whether or not what we’re playing is the latest version. The note is sent from the Isolate.

For example, Chrome on Android won’t reload isolates (webworkers) unless forced by `javascript:location.reload(true)` magic. Without a note about version, it’s easy to forget about this and test an old version of the game thinking we’re looking at the new one.