Fix bug with getRollReason
Previously, we threw for getRollReason for writer’s input actions under the assumptions that you never want to call that function.

It was a mistake: I was thinking about _reroll_ (which shouldn’t be there, yet). But we do want to roll for these, and shouldn’t throw.

This also reverts the code path that didn’t show slotmachine for the player when there was no reroll reason.