SimpleAction’s success closure now gets `self`
Before, we provided the `movePlayer` function, but that’s not scalable. Now, when we need to provide such functions, they can be accessed by `self.movePlayer` et al.

This also fixes bug which moved player to mountain_pass without showing the description. Now, `moveActor` will automatically show description, and is the canonical way to move around. Before, `TakeExit` was taking care of this separately, and `moveActor`/`movePlayer` didn’t.