Implement `[[CODE]]` blocks and world.actionIsBeingPerformed
Code blocks allow custom code in descriptions. Currently only supported in Room description and short_description. This is the chose alternative to [[IF]] in text. https://trello.com/c/tyyguP0T/14-code-syntax

`WorldState.actionIsBeingPerformed()` is an equivalent to `WorldState.actionHasBeenPerformedSuccessfully` for the action that is currently being applied. This matters for when a Room description needs to know which action triggered it, so that it can modify its output (for example, the player can arrive on foot or on horseback).