Report Briana’s near-death
A bug in the code meant that when Briana got her final blow in a FightSituation (hitpoints == 0), the “she goes down” report wouldn’t be used. This was because reportPain was called with the `enemy` object from _before_ the hitpoints were removed.

Fixed by providing the updated enemy object to reportPain and killHumanoid.