Do not report “THE END” at the end
Author can do that either in a ROOM description, in an ACTION description, or in .egb.

This gives the author more control.

We still report ‘You died.’ because author currently has no way of providing this particular report to the user. This could be remedied by allowing an interrupt, or by providing an enclosing “Life” situation (above all other situations, including RoomRoaming).