Introduce Action.isImplicit and Choice.isImplicit
Instead of the hacky way of marking actions as “automatic” by having the `command` be an empty string, we now have an explicit isAutomatic in both the Choice and the Action itself.

This removes Choice.isAutomatic (which was just a shorthand way of checking if `command` is and empty string).