Create weapons from fists
Instead of just creating new fists when actor is barehanded, the system walks the actor’s anatomy and picks a body part that can act as a weapon. Then it creates a new item/weapon backed by the body part’s `damageCapability`.

This allows things like clawed or tailed humanoids. In the longer term, this paves the way of rats, giant spiders, etc., where anatomy is an important source of weapons.

Currently, the system will throw when actor loses a weapon and doesn’t have a usable bodypart. We’ll need to make `currentWeapon` nullable and allow actors to just not have any weapon at all.