Switch to build.yaml-based source-generation
This is WIP. `built_value` works but other things don’t.

Situation is now fully abstract, as is Entity. All inherited behavior is in EntityBehavior and SituationBaseBehavior, respectively. This is because `built_value` doesn’t generate code for fields that have been defined as final in the super class. They must be defined as getters all the way up, it seems.

Also, there was a bug with many classes that *extended* `Built<T,S>` instead of implementing it. This is caught by the newer built_value.